Sales Order view modified

+++ NBL/Areas/Sales/BLL/Contracts/IInvoiceManager.cs
@@ -18,9 +18,9 @@ namespace NBL.Areas.Sales.BLL.Contracts
         IEnumerable<Invoice> GetAllInvoicedOrdersByUserId(int invoiceByUserId);
         IEnumerable<Invoice> GetInvoicedRefferencesByClientId(int clientId);
         IEnumerable<Invoice> GetAllInvoicedOrdersByCompanyId(int companyId);
-        IEnumerable<InvoiceDetails> GetInvoicedOrderDetailsByInvoiceId(int invoiceId);
+        IEnumerable<InvoiceDetails> GetInvoicedOrderDetailsByInvoiceId(long invoiceId);
         IEnumerable<InvoiceDetails> GetInvoicedOrderDetailsByInvoiceRef(string invoiceRef);
-        Invoice GetInvoicedOrderByInvoiceId(int invoiceId);
+        Invoice GetInvoicedOrderByInvoiceId(long invoiceId);
         ICollection<ViewProduct> GetDeliveredProductsByInvoiceRef(string invoiceRef);
         ICollection<Invoice> GetInvoicedOrdersByCompanyIdAndDate(int companyId, DateTime date);
         ICollection<Invoice> GetLatestInvoicedOrdersByDistributionPoint(int distributionPointId);
